+++ e2e/e2e-helper.js
@@ -249,6 +249,9 @@ export default class Helper {
   addComponent(filePaths: string = path.normalize("bar/foo.js"), cwd = this.localScopePath) {
     return this.runCmd(`bit add ${filePaths}`, cwd);
   }
+  untrackComponent(id: string='', cwd = this.localScopePath) {
+    return this.runCmd(`bit untrack ${id}`, cwd);
+  }
 
   addComponentWithOptions(filePaths: string = 'bar/foo.js', options:? Object) {
     const value = Object.keys(options).map(key => `-${key} ${options[key]}`).join(' ');