+++ src/api/scope/lib/cat-object.js
@@ -1,12 +1,12 @@
 /** @flow */
 import { loadScope } from '../../../scope';
 
-export default function catObject(hash: string) {
+export default function catObject(hash: string, pretty: boolean) {
   return loadScope().then((scope) => {
     return scope.getObject(hash)
       .then((object) => {
         if (!object) return 'object not found';
-        return object.toBuffer();
+        return object.toBuffer(pretty);
       });
   });
 }