+++ src/utils/index.js
@@ -45,6 +45,7 @@ import objectToStringifiedTupleArray from './object-to-stringified-tuple-array';
 import deflate from './zlib-deflate';
 import sha1 from './encryption/sha1';
 import isString from './string/is-string';
+import getStringifyArgs from './string/get-stringify-args';
 import isNumeric from './number/is-numeric';
 import diff from './array/diff';
 import filterObject from './filter-object';
@@ -59,6 +60,7 @@ import splitBy from './array/split-by';
 import outputFile from './fs-output-file';
 import getLatestVersionNumber from './resolveLatestVersion';
 import calculateFileInfo from './fs/file-info';
+import getWithoutExt from './fs/fs-no-ext';
 import outputJsonFile from './fs-output-json-sync';
 import { pathNormalizeToLinux, pathJoinLinux, pathRelative, pathResolve, pathJoinOs } from './path';
 
@@ -73,6 +75,7 @@ export {
   rmDir,
   filterObject,
   isString,
+  getStringifyArgs,
   isNumeric,
   inflate,
   diff,
@@ -124,6 +127,7 @@ export {
   bufferFrom,
   getLatestVersionNumber,
   calculateFileInfo,
+  getWithoutExt,
   pathNormalizeToLinux,
   pathJoinLinux,
   pathRelative,